lab-2/migrations/versions/0001_create_notes_table.py
"""create notes table

Revision ID: 0001_create_notes_table
Revises:
Create Date: 2026-02-28 23:45:00.000000

"""

from alembic import op
import sqlalchemy as sa


revision = "0001_create_notes_table"
down_revision = None
branch_labels = None
depends_on = None


def upgrade():
    op.create_table(
        "notes",
        sa.Column("id", sa.Integer(), nullable=False),
        sa.Column("title", sa.String(length=120), nullable=False),
        sa.Column("content", sa.Text(), nullable=False),
        sa.Column("created_at", sa.DateTime(timezone=True), nullable=False),
        sa.Column("updated_at", sa.DateTime(timezone=True), nullable=False),
        sa.PrimaryKeyConstraint("id"),
    )


def downgrade():
    op.drop_table("notes")
