lab-2/requirements.txt
Flask==3.1.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
gunicorn==23.0.0
psycopg[binary]==3.2.6
